Uploads Manager: Move to core\files
